Tabela de configuração dos perfis que poderão criar issues criada

+++ hook.php
@@ -27,9 +27,21 @@ function plugin_gitlabintegration_install(){
 					FOREIGN KEY (`ticket_id`) REFERENCES `glpi_tickets` (`id`)";
 	    $DB->queryOrDie($query, $DB->error());
 	}
-	// To be called for each task the plugin manage
-   	// task in class
-   	CronTask::Register('PluginExampleExample', 'Sample', DAY_TIMESTAMP, ['param' => 50]);
+
+	//Create table glpi_plugin_gitlab_integration only if it does not exists yet!
+	if (!$DB->tableExists('glpi_plugin_profiles_users')) {
+	    $query = "CREATE TABLE `glpi_plugin_profiles_users` (
+				   `id` INT(11) NOT NULL AUTO_INCREMENT,
+				   `profile_id` INT(11) NOT NULL,
+				   PRIMARY KEY  (`id`)
+				) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci";
+	    $DB->queryOrDie($query, $DB->error());
+
+	    $query = "ALTER TABLE `glpi_plugin_profiles_users` 
+	                ADD CONSTRAINT `fk_gitlab_profile` 
+					FOREIGN KEY (`profile_id`) REFERENCES `glpi_profiles` (`id`)";
+	    $DB->queryOrDie($query, $DB->error());
+	}
  
 	return true;
 }
@@ -52,8 +64,15 @@ function plugin_gitlabintegration_uninstall(){
 		$notif->delete($data);
 	}
 
-	$drop_count = "DROP TABLE glpi_plugin_gitlab_integration";
-	$DB->query($drop_count); 	
+	if ($DB->tableExists('glpi_plugin_gitlab_integration')) {
+		$drop_count = "DROP TABLE glpi_plugin_gitlab_integration";
+		$DB->query($drop_count); 
+	}
+	
+	if ($DB->tableExists('glpi_plugin_profiles_users')) {
+		$drop_count = "DROP TABLE glpi_plugin_profiles_users";
+		$DB->query($drop_count);
+	} 
 	
 	return true;
 }
